requirements.txt
tqdm~=4.64.1
pyyaml~=6.0
ultralytics~=8.0.42
opencv-python~=4.7.0.68
numpy~=1.24.2
torch~=1.13.1
pillow~=9.4.0
setuptools~=65.6.3
thop~=0.1.1-2209072238
pandas~=1.5.3
requests~=2.28.2
torchvision~=0.14.1
build~=0.10.0
psutil~=5.9.4
clearml~=1.9.1
matplotlib~=3.7.0
scipy~=1.10.1